requirements.txt
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bidict==0.22.1
boto3==1.26.45
botocore==1.29.45
cachetools==5.3.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
eventlet==0.33.2
Flask==2.2.2
Flask-SocketIO==5.3.2
google-auth==2.16.0
greenlet==2.0.1
h11==0.13.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
numpy==1.23.3
outcome==1.2.0
pandas==1.4.4
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
PyJWT==2.6.0
pymongo==4.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.2.1
redis==4.5.1
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
